% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_simple_outlier.R
\name{is_simple_outlier}
\alias{is_simple_outlier}
\title{Determine outliers.Well suited in dplyr pipes
TODO: Inlcude complex data such as time series outlier}
\usage{
is_simple_outlier(x, l = 1.5)
}
\arguments{
\item{x}{Numeric column}

\item{l}{Lower limit and higher limits}
}
\value{
Logical vector indication outliers
}
\description{
Determine outliers.Well suited in dplyr pipes
TODO: Inlcude complex data such as time series outlier
}
\examples{
\dontrun{
mutate_if(is.numeric, funs(outlier = is_simple_outlier(.,1.5)))
}
}
